[SHELL32] Optimize change notification
